Widen site_id to UInt32 so an instance can hold more than 65535 sites

+++ server/src/api/analytics/runCustomQuery.ts
@@ -60,7 +60,7 @@ export async function runCustomQuery(
     WITH scoped_events AS (
       SELECT *
       FROM events
-      PREWHERE site_id IN {siteIds:Array(UInt16)}
+      PREWHERE site_id IN {siteIds:Array(UInt32)}
     )
     SELECT *
     FROM (
